chore(functions): #1213 refactored cloud functions and exports
