+++ functions/src/client/firebase-admin.ts
@@ -1,0 +1,3 @@
+import * as admin from 'firebase-admin';
+
+export const FirebaseAdmin: admin.app.App = admin.initializeApp();
