+++ functions/src/index.ts
@@ -1,7 +1,12 @@
-import * as admin from 'firebase-admin';
+import { getRepositoryInfo, RepositoryInfoInput } from './repository/info';
+import * as functions from 'firebase-functions';
+import { CallableContext } from 'firebase-functions/lib/providers/https';
 
-export const FirebaseAdmin: admin.app.App = admin.initializeApp();
+import { ReposInput, getUserRepos } from './user/repos';
+import { getUserEvents, EventsInput } from './user/events';
 
-export * from './user/repos';
-export * from './user/events';
-export * from './repository/info';
+declare type HttpsFunction = functions.HttpsFunction;
+
+export const findAllUserRepositories: HttpsFunction = functions.https.onCall((input: ReposInput, context: CallableContext) => getUserRepos(input.token, context.auth.uid));
+export const findAllUserEvents: HttpsFunction = functions.https.onCall((input: EventsInput, context: CallableContext) => getUserEvents(input.token, context.auth.uid, input.username));
+export const findRepositoryInfo: HttpsFunction = functions.https.onCall((input: RepositoryInfoInput, context: CallableContext) => getRepositoryInfo(input.token, input.fullName));
