+++ functions/src/user/events.ts
@@ -1,26 +1,24 @@
-import { FirebaseAdmin } from './../index';
-import * as functions from 'firebase-functions';
-import { CallableContext } from 'firebase-functions/lib/providers/https';
+import { FirebaseAdmin } from './../client/firebase-admin';
 
 import { GitHubEventInput, GitHubEventMapper, GitHubEventModel } from '../mappers/github/index.mapper';
 import { GitHubClient } from './../client/github';
 
-interface Input {
+export interface EventsInput {
     token: string;
     username: string;
 }
 
-export const getUserEvents: functions.HttpsFunction = functions.https.onCall(async (input: Input, context: CallableContext) => {
-    const events: GitHubEventInput[] = await GitHubClient<GitHubEventInput[]>(`/users/${input.username}/events`, input.token);
+export const getUserEvents: any = async (token: string, uid: string, username: string) => {
+    const events: GitHubEventInput[] = await GitHubClient<GitHubEventInput[]>(`/users/${username}/events`, token);
     const mappedEvents: GitHubEventModel[] = events.map((event: GitHubEventInput) => GitHubEventMapper.import(event));
 
     await FirebaseAdmin
         .firestore()
         .collection('users')
-        .doc(context.auth.uid)
+        .doc(uid)
         .set({
             activity: mappedEvents,
         }, { merge: true });
 
     return mappedEvents;
-});
+};
