+++ functions/src/user/stats.ts
@@ -1,7 +1,7 @@
 import * as admin from 'firebase-admin';
 import { Change, EventContext, firestore } from 'firebase-functions';
+import { FirebaseAdmin } from './../client/firebase-admin';
 
-import { FirebaseAdmin } from './../index';
 import { getUserRepos } from './repos';
 import { GitHubUserStatsModel } from './../mappers/github/user.mapper';
